requirements_dev.txt
bentoml==1.0.10
joblib==1.2.0
pip-chill==1.0.1
torchvisionpip3 install starlette
tqdm==4.64.1
wincertstore==0.2
dvc
mlflow
ipykernel
numpy 
pandas
seaborn

pytest==7.1.3
tox==3.25.1
flake8==5.0.4
mypy==0.971
black==22.8.0


-e .